Rename all-sky response top FullDetectorResponse. Response for a particular direction to simply DetectorResponse
